Doplnění testovacích dat a oprava filtrů

+++ Core/AttachmentModel.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
+using System.IO;
 
 namespace Core
 {
@@ -12,6 +13,8 @@ namespace Core
         /// Uchovává databázový kontext.
         /// </summary>
         protected DbPublicationEntities context;
+
+        private const string DATA_DIRECTORY = "data/";
 
         /// <summary>
         /// Vytvoří instanci správce.
@@ -45,8 +48,12 @@ namespace Core
         /// </summary>
         /// <param name="publicationId">ID publikace</param>
         /// <param name="attachment">údaje o příloze</param>
-        public void AddAttachmentToPublication(Publication publication, Attachment attachment)
+        public void AddAttachmentToPublication(Publication publication, string srcFileName)
         {
+            Attachment attachment = new Attachment();
+            attachment.Path = DATA_DIRECTORY + Path.GetFileName(srcFileName);
+            File.Copy(srcFileName, attachment.Path);
+            
             attachment.Publication = publication;
             List<Attachment> attachmentList = GetAttachmentsByPublication(publication);
             // určení ID nové přílohy
@@ -55,13 +62,13 @@ namespace Core
             context.Attachment.Add(attachment);
             context.SaveChanges();
         }
-        
+
         /// <summary>
-        /// Odebere z publikace existující přílohu.
+        /// Zkopíruje existující přílohu publikace.
         /// </summary>
         /// <param name="publicationId">ID publikace</param>
         /// <param name="id">ID přílohy</param>
-        public void RemoveAttachmentFromPublication(Publication publication, int id)
+        public void CopyAttachmentOfPublication(Publication publication, string destFileName, int id)
         {
             Attachment attachment = context.Attachment.Find(publication.Id, id);
 
@@ -71,9 +78,29 @@ namespace Core
                     "Příloha s id {0} u publikace s id {1} neexistuje.", id, publication.Id));
             }
 
+            File.Copy(attachment.Path, destFileName);
+        }
+
+        /// <summary>
+        /// Odebere z publikace existující přílohu.
+        /// </summary>
+        /// <param name="publicationId">ID publikace</param>
+        /// <param name="id">ID přílohy</param>
+        public void RemoveAttachmentFromPublication(Publication publication, int id)
+        {
+            Attachment attachment = context.Attachment.Find(publication.Id, id);
+            
+            if (attachment == null)
+            {
+                throw new AttachmentException(string.Format(
+                    "Příloha s id {0} u publikace s id {1} neexistuje.", id, publication.Id));
+            }
+
             publication.Attachment.Remove(attachment);
             context.Attachment.Remove(attachment);
             context.SaveChanges();
+
+            File.Delete(attachment.Path);
         }
     }
 }
